fix index page
