+++ app/commands/decidim/anonymous_codes/admin/create_tokens.rb
@@ -12,11 +12,7 @@ module Decidim
         def call
           return broadcast(:invalid) if form.invalid?
 
-          transaction do
-            form.num_tokens.times do
-              create_token!
-            end
-          end
+          CreateTokensJob.perform_later(code_group, form.num_tokens)
 
           broadcast(:ok)
         end
@@ -24,19 +20,6 @@ module Decidim
         private
 
         attr_reader :form, :code_group
-
-        def create_token!
-          token = new_token while token.blank?
-          token.save!
-        end
-
-        def new_token
-          token = Decidim::AnonymousCodes::Token.new(
-            token: Decidim::AnonymousCodes.token_generator,
-            group: code_group
-          )
-          token if token.valid?
-        end
       end
     end
   end
