+++ app/views/decidim/anonymous_codes/admin/code_groups/_form.html.erb
@@ -1,6 +1,8 @@
 <div class="card">
   <div class="card-divider">
-    <h2 class="card-title"><%= title %></h2>
+    <h2 class="card-title"><%= title %>
+      <a class="button tiny button--title hollow" href="<%= code_groups_path %>"><%= t("codes.index.back", scope: "decidim.anonymous_codes.admin") %></a>
+    </h2>
   </div>
 
   <div class="card-section">
