+++ app/views/decidim/anonymous_codes/admin/codes/_form.html.erb
@@ -1,6 +1,9 @@
 <div class="card">
   <div class="card-divider">
-    <h2 class="card-title"><%= title %></h2>
+    <h2 class="card-title"><%= title %>
+
+      <a class="button tiny button--title hollow" href="<%= code_group_codes_path(code_group) %>"><%= t("codes.index.back_to_codes", scope: "decidim.anonymous_codes.admin") %></a>
+    </h2>
   </div>
 
   <div class="card-section">
